backend/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask-cors
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
paho-mqtt==2.1.0
psycopg2-binary==2.9.9
sqlalchemy
alembic
Werkzeug==3.0.4
pytest
pytest-cov==4.1.0
pytest-html
